requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.12.0
whois==0.9.27
aiohttp==3.9.1
requests==2.31.0
jinja2==3.1.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0 
Jinja2==3.1.2
nltk
starlette
python-multipart
mailersend==0.2.5